backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.11.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
cloudinary==1.42.2
python-dotenv==1.0.1
pydantic[email]==2.10.6
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
